+++ src/calendar/components/Navbar.jsx
@@ -2,12 +2,12 @@ export const Navbar = () => {
   return (
     <div className='navbar navbar-dark bg-dark mb-4 px-4'>
       <span className='navbar-brand'>
-        <i className='fas fa-calendar-alt'></i>
-        &nbsp; Jonathan
+        <i className='fas fa-calendar-alt me-2'></i>
+        Jonathan
       </span>
       <button className='btn btn-outline-danger'>
-        <i className='fas fa-sign-out-alt'></i>
-        <span>&nbsp; Salir</span>
+        <i className='fas fa-sign-out-alt me-2'></i>
+        <span>Salir</span>
       </button>
     </div>
   );